dashboard: do not try to create rgw system user on secondary cluster
When deploying/upgrading a secondary rgw multisite cluster, this task
fails with following message:

```
Please run the command on master zone. Performing this operation on non-master zone leads to inconsistent metadata between zones
Are you sure you want to go ahead? (requires --yes-i-really-mean-it)
```

Let's skip this task when the node is a secondary one.

Closes: https://bugzilla.redhat.com/show_bug.cgi?id=1794351

Signed-off-by: Guillaume Abrioux <gabrioux@redhat.com>