Don't create RGW pools on upgrade
Skip the RGW pool creation tasks on upgrade. The same is already
done for OpenStack pools.

When upgrading from an older version which may have created a different
set of pools or pools with different parameters, creating the pools
could create unneeded new pools or cause pool parameters to change.

Signed-off-by: Gaudenz Steinlin <gaudenz.steinlin@cloudscale.ch>