[WIP] Fix delegation when facts are computed in rolling_update context
This change is trying to fix an issue that occurs when rolling_update
playbook is run: in particular the following error is reported:

 FAILED! => {"msg": "list object has no element 0"}

and this patch is just an attempt to fix it.
The delegate_to should be run on the current selcted monitor, or
it will fail looking for a container which is running on a different
host.

Closes: https://bugzilla.redhat.com/1877426
Signed-off-by: Francesco Pantano <fpantano@redhat.com>