Disable redirect mechanism in a ssl+haproxy context
When ssl is enabled and haproxy is configured (because the vip is
provided only when we have haproxy entries), according to the doc
[1][2] we want to disable the redirection behaviour to prevent
situations when internal (unresolvable) URL’s are published to the
frontend client.
By doing this haproxy is able to respond with a HTTP error (503),
sending the user to the active instance.

[1] https://docs.ceph.com/en/latest/mgr/dashboard/#disable-the-redirection
[2] https://docs.ceph.com/en/latest/mgr/dashboard/#haproxy-example-configuration

Signed-off-by: fmount <fpantano@redhat.com>